testing

+++ midterm1/main.c
@@ -16,7 +16,7 @@ extern long * manager();
 
 int main()
 {
-    long * arrayPtr;
+    double * arrayPtr;
     printf("\e[1;33mWelcome to the Great Reverse by Victor Mercado. \n\e[33m");
     arrayPtr = manager();
     printf("\e[1;33mThe driver function received an array whose first 3 numbers are: \n\e[0m");
